requirements.txt
discord.py
gTTS